added styles file
